requirements.txt
fastmcp
supabase
python-dotenv
langchain
langchain-google-genai
google-generativeai
pytest
pytest-asyncio
pydantic